+++ django_gcp/storage/__init__.py
@@ -1,7 +1,8 @@
 # TODO: Refactor tests to import members directly then remove this module export
 from . import gcloud
-from .blob_utils import BlobFieldMixin, get_blob, get_blob_name, get_path, get_signed_url, upload_blob
+from .blob_utils import BlobFieldMixin, get_blob, get_blob_name, get_path, get_signed_url
 from .gcloud import GoogleCloudFile, GoogleCloudMediaStorage, GoogleCloudStaticStorage, GoogleCloudStorage
+from .operations import upload_blob
 
 __all__ = [
     "GoogleCloudStorage",
